+++ packages/core/__tests__/context.tsx
@@ -1,7 +1,13 @@
 import { renderHook } from '@testing-library/react-hooks'
+import type { Map } from 'leaflet'
 import React, { type ReactNode } from 'react'
 
-import { LeafletProvider, useLeafletContext } from '../src'
+import {
+  CONTEXT_VERSION,
+  LeafletProvider,
+  createLeafletContext,
+  useLeafletContext,
+} from '../src'
 
 export function createWrapper(context) {
   return function Wrapper({ children }: { children: ReactNode }) {
@@ -10,6 +16,14 @@ export function createWrapper(context) {
 }
 
 describe('context', () => {
+  test('createLeafletContext() creates the context object', () => {
+    const map = {} as unknown as Map
+    expect(createLeafletContext(map)).toEqual({
+      __version: CONTEXT_VERSION,
+      map,
+    })
+  })
+
   test('useLeafletContext() throws an error when there is no provider', () => {
     expect(() => {
       const { result } = renderHook(() => useLeafletContext())